Fix URP serialization issue
