+++ source/com.unity.cluster-display.graphics/Runtime/Camera/InjectionPointRenderFeature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae01727733165294ebe5c1f1866bca22
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
