+++ source/com.unity.cluster-display.graphics/Runtime/Camera/UrpPresenter.cs
@@ -8,41 +8,6 @@ namespace Unity.ClusterDisplay.Graphics
 {
     class UrpPresenter : IPresenter
     {
-        /// <summary>
-        /// A render pass whose purpose is to invoke an event at the desired stage of rendering.
-        /// </summary>
-        class InjectionPointRenderPass : ScriptableRenderPass
-        {
-            public static event Action<ScriptableRenderContext, RenderingData> ExecuteRender = delegate {};
-
-            public override void Execute(ScriptableRenderContext context, ref RenderingData renderingData)
-            {
-                ExecuteRender.Invoke(context, renderingData);
-            }
-        }
-
-        /// <summary>
-        /// A render feature whose purpose is to provide an event invoked at a given rendering stage.
-        /// Meant to abstract away the render feature mechanism and allow for simple graphics code injection.
-        /// </summary>
-        internal class InjectionPointRenderFeature : ScriptableRendererFeature
-        {
-            InjectionPointRenderPass m_Pass;
-
-            public override void Create()
-            {
-                m_Pass = new InjectionPointRenderPass
-                {
-                    renderPassEvent = RenderPassEvent.AfterRendering,
-                };
-            }
-
-            public override void AddRenderPasses(ScriptableRenderer renderer, ref RenderingData renderingData)
-            {
-                renderer.EnqueuePass(m_Pass);
-            }
-        }
-        
         const string k_CommandBufferName = "Present To Screen";
 
         Camera m_Camera;
